exemplos-extras/exemplo-ecommerce/ecommerce-etl/requirements.txt
pandas
sqlalchemy
google-cloud-bigquery
faker
numpy